requirements.txt
numpy==1.16.3
pandas==0.24.2
jsonschema==3.0.1
requests==2.22
Flask==1.0.2
pymongo==3.8.0
Flask_Bcrypt==0.7.1
Flask_JWT_Extended==3.18.2
xlrd==1.2.0
coloredlogs==10.0
Flask-Script==2.0.6
termcolor==1.1.0
watchdog==0.9.0
pathlib==1.0.1
uwsgi==2.0.18
mt-940==4.15.0
alpha_vantage==2.1.0
schedule==0.6.0
python-dateutil==2.8.0
quandl==3.4.8
python-dotenv==0.10.3
